Introduce temporary keys
A temporary key can be used in lieu
of the device but has a short expiration time. These keys are used as a workaround for WebAuthn
needing two users interactions: one for "create" and one for "sign". So instead we only "create"
and instead authenticate the user with a temporary key for their first visit.

Note: we link the temporary keys to a device (as opposed to an anchor) so that we can make sure
the temporary key is dropped if the device itself is removed. This ensures the temporary key is
no more powerful than the device, i.e. if the user decides to remove the device right after
registration, then the the temporary key cannot be used to authenticate (similary to how the
browser session key pair cannot be used to authenticate if the actual delegated WebAuthn device is
removed)